Migrate 013-quarkus-oidc-restclient module from Beefy TS
This module has been migrated into the oidc-client-extended module as it requires new dependencies like reactive and restclient extensions.
I also renamed the keycloak-oidc-client to keycloak-oidc-client-basic so it's clear the scope of each one.

Moreover, I refactor the solution to run all the coverage in one go, so we don't need to deploy a keycloak to test every rest client path.

Finally, I only mark the test RestClientInPingPongResourcesIT to be run on OpenShift on purpose as it's the only relevant coverage here.