Fix test: PostgresPoolIT.checkDbPoolTurnover
The problem was caused by calling to the Reactive Route using RESTAssured instead of using the WebClient sockets. 
Calling using RestAssured was making the call being ended before the response (the number of active connections) was indeed returned.
The rest of changes are cosmetic, so this test looks similar to the one in Beefy.
I did check that now it's working on Native and JVM.