QUARKUS-1676 Add tests for 'MAIF - mTLS with OIDC Server'
This commit add tests for a mutual TLS with OIDC as suggested by [a test plan](https://github.com/quarkus-qe/quarkus-test-plans/pull/76). There are tests for JKS/PKCS12 keystore types, a fail test when a keystore file is completely missing, test with Keycloak and RH SSO 7.3/7.5 in role of a OIDC server and test with keystore file whose file type can't be detected by a file extension.
refactor according to review comments and add a new test