Fix "unexpected end of stream" error.
We check that MariaDB container started, by looking into log and waiting
for special lines. Before that change, we were looking for a line "ready for connections",
which was duplicated and its first appearance was right before a reload
of the container, so in some cases we were trying to connect to
non-working DBMS. Now we use a unique line.