Split bouncycastle module into two modules
Rest-assurance client is not working with BCFIPS certificates, swap to Vertx HTTP client.

Also, we have to split the module in two because QuarkusScenarios doesn't deal well with bouncyCastle Jsse
so we used QuarkusTest instead. But...quarkusTest doesn't support "pom dependencies injection for a single scenario", so this is why we end with two different modules.