Test MessageBodyWriter receives right accepted Media Type with Response
Verifies [Quarkus Issue 22119](https://github.com/quarkusio/quarkus/issues/22119). According to JAX-RS standard, if request specifies accepted Media type, the type should be passed to `javax.ws.rs.ext.MessageBodyWriter`, however for use case when `javax.ws.rs.core.Response` is returned, that was only fixed in Quarkus 2.8.1 and backported to 2.7.6. Previously, if multiple media types were defined in `@Produces`, the first one was taken. This test verifies that a value of `Accept` HTTP header is same as the media type passed to `MessageBodyWriter` when endpoint response type is `Response`.