Fix OpenShiftGraphQLTelemetryIT - filter Jaeger traces by operation name
Currently traces are filtered by "operationName" which is not valid filter parameter and operations are not filtered by this. That causes problem in OpenShift where there are other calls (other operations with different names, e.g. "HTTP GET" - probably probe check). Problem can be easily reproduced locally on GraphQLTelemetryIT when you call GET to '/' couple of times, which leads to test failure.