Cover separate management interface
Now Quarkus can expose endpoints for metrics and health
on a separate network interface.
We verify, that this is supported, and we can use custom port, switch
HTTPS on and off and change api path.
This requires supports on the side of the FW and enabling one of the
disabled tests.

Test plans for the change: https://github.com/quarkus-qe/quarkus-test-plans/blob/main/QUARKUS-2738.md
Related links:
https://github.com/quarkusio/quarkus/pull/30506
https://issues.redhat.com/browse/QUARKUS-2738