Fix analytics disabled by CI
Build-time analytics can be disabled by `CI` environment variable. This
fix explicitly sets the value of the variable to `false` for both
surefire and failsafe executions in the analytics module, so that it
gets executed in every environment.
The module itself works around the reasons this feature has been added
to Quarkus, which is the fact that the initial analytics user prompt
cannot be acted upon in CI environments.

See https://github.com/quarkusio/quarkus/pull/35456.