Add coverage for ArC fix when a framework bean uses the decorator
This adding the coverage for https://issues.redhat.com/browse/QUARKUS-5178
It's tested with Quarkiverse extension as the reproducer attached in issue
and it was only scenario which I was able to reproduce.

To test this you need external lib which contains class (lib_class) implementing the interface.
This interface is also aplied on class anotated as decorator in Quarkus app when the bean of lib_class is injected at the same time.

This error happen randomly in dev/test mode.

The Quarkus PR is https://github.com/quarkusio/quarkus/pull/43245