Fix wrong creation and resolution of certificate path in `BearerFilesystemIT`
The `Path.of`  failed to create due to g`etResource().getPath()`
which started always with `\` and created invalid path on windows.

The `tokenFilePath` on windows added `\` at the end of tmp directory and when construction `tokenFilePath`
the another separator was added causing that two separators were placed after each other.
When the log were searched they expect the path `C:\<path>\\abcdefgh.token` but log contain `C:\<path>\abcdefgh.token`